test_utils: proven_tx module
