+++ block-producer/src/block_builder/errors.rs
@@ -2,7 +2,7 @@ use miden_objects::accounts::AccountId;
 use miden_vm::{crypto::MerkleError, ExecutionError};
 use thiserror::Error;
 
-use crate::store::{BlockInputsError, ApplyBlockError};
+use crate::store::{ApplyBlockError, BlockInputsError};
 
 #[derive(Debug, Error, PartialEq, Eq)]
 pub enum BuildBlockError {
